Rename ping_one_connection to pingone_connection
