+++ examples/resources/pingfederate_pingone_connection/import.sh
@@ -1,0 +1,1 @@
+terraform import pingfederate_pingone_connection.pingOneConnection connectionId
