+++ examples/resources/pingfederate_pingone_connection/resource.tf
@@ -1,0 +1,4 @@
+resource "pingfederate_pingone_connection" "example" {
+  name       = "My PingOne Environment"
+  credential = var.pingone_connection_credential
+}
