+++ internal/provider/provider.go
@@ -781,7 +781,8 @@ func (p *pingfederateProvider) Resources(_ context.Context) []func() resource.Re
 		oauthtokenexchangegeneratorsettings.OauthTokenExchangeGeneratorSettingsResource,
 		oauthtokenexchangetokengeneratormapping.OauthTokenExchangeTokenGeneratorMappingResource,
 		passwordcredentialvalidator.PasswordCredentialValidatorResource,
-		pingoneconnection.PingOneConnectionResource,
+		pingoneconnection.DeprecatedPingOneConnectionResource,
+		pingoneconnection.PingoneConnectionResource,
 		protocolmetadatalifetimesettings.ProtocolMetadataLifetimeSettingsResource,
 		redirectvalidation.RedirectValidationResource,
 		serversettings.ServerSettingsResource,
